cli installer: fixes from Martin Langhoff (OLPC XS)
This includes the following commits:

0a0b18858447f0a9925f290759333d055261bbfe
9a208bd63df739fd4105ae2cdb5c6cb0bf63d661
189c731aed2ebb959107536de20ddc8c660e9ea2
1ab976d3b675134da8d3694b3d1edaac7b8dcfbb
49d9375562d808a7e6d50f7e1ce84d76eeaf91c5

+++ admin/cliupgrade.php
@@ -723,7 +723,7 @@ if ( file_exists(dirname(dirname(__FILE__)) . '/config.php')) {
 
     /// Check settings in config.php
 
-    $dirroot = dirname(realpath("../index.php"));
+    $dirroot = realpath(dirname(dirname(__FILE__)));
     if (!empty($dirroot) and $dirroot != $CFG->dirroot) {
         console_write(STDERR,"Please fix your settings in config.php:
               \nYou have:
@@ -901,6 +901,23 @@ if ( file_exists(dirname(dirname(__FILE__)) . '/config.php')) {
     if ($CFG->version) {
         if ($version > $CFG->version) {  // upgrade
 
+            // Upgrades: installations made with early versions of
+            // cliupgrade.php may be missing $CFG->release - and the
+            // environment checks need id. Provide an approx release
+            // string, based on the
+            if (empty($CFG->release)) {
+                if ((int)$CFG->version >= 2007101509) {
+                    $release_to_upgrade = '1.9';
+                } elseif ((int)$CFG->version >= 2007021501) {
+                    $release_to_upgrade = '1.8';
+                }
+                if (!empty($release_to_upgrade)) {
+                    set_config('release', $release_to_upgrade);
+                    $CFG->release = $release_to_upgrade;
+                }
+            }
+
+
             /// If the database is not already Unicode then we do not allow upgrading!
             /// Instead, we print an error telling them to upgrade to 1.7 first.  MDL-6857
             if (empty($CFG->unicodedb)) {
@@ -1007,12 +1024,12 @@ if ( file_exists(dirname(dirname(__FILE__)) . '/config.php')) {
                 if (!stats_upgrade_for_roles_wrapper()) {
                     notify('Couldn\'t upgrade the stats tables to use the new roles system');
                 }
-                if (set_config("version", $version)) {
+                if (set_config("version", $version) && set_config("release", $release)) {
                     remove_dir($CFG->dataroot . '/cache', true); // flush cache
                     notify($strdatabasesuccess, "green");
                     /// print_continue("upgradesettings.php");
                 } else {
-                    console_write(STDERR,'Upgrade failed!  (Could not update version in config table)','',false);
+                    console_write(STDERR,'Upgrade failed!  (Could not update version or release in config table)','',false);
                 }
                 /// Main upgrade not success
             } else {
@@ -1029,6 +1046,9 @@ if ( file_exists(dirname(dirname(__FILE__)) . '/config.php')) {
     } else {
         if (!set_config("version", $version)) {
             console_write(STDERR,"A problem occurred inserting current version into databases",'',false);
+        }
+        if (!set_config("release", $release)) {
+            console_write(STDERR,"A problem occurred inserting current release into databases",'',false);
         }
     }
 
@@ -1115,10 +1135,19 @@ if ( file_exists(dirname(dirname(__FILE__)) . '/config.php')) {
 
     /// Check all message output plugins and upgrade if necessary
     if ( $verbose > CLI_NO ) {
-        print_heading(get_string('upgradingmessageoutputpluggin','install'),'',1);
+        print_heading(get_string('upgradingmessageoutputplugin','install'),'',1);
     }
     upgrade_plugins('message','message/output',''); // Don't return anywhere
 
+    if ( $verbose > CLI_NO ) {
+        print_heading(get_string('upgradingcoursereportplugin','install'),'',1);
+    }
+    upgrade_plugins('coursereport', 'course/report','');
+
+    if ( $verbose > CLI_NO ) {
+        print_heading(get_string('upgradingadminreportplugin','install'),'',1);
+    }
+    upgrade_plugins('report', (empty($CFG->admin) ? 'admin' : $CFG->admin) .'/report', '');
 
     /// just make sure upgrade logging is properly terminated
     upgrade_log_finish();
@@ -1302,6 +1331,14 @@ if ( file_exists(dirname(dirname(__FILE__)) . '/config.php')) {
 
         create_admin_user($user);
     }
+
+    // Set all the settings to their defaults. We need
+    // to fake admin rights for this :-(
+    $admuser = get_admin();
+    $USER = get_complete_user_data('id', $admuser->id);
+    complete_user_login($USER);
+    admin_new_settings_to_default();
+
     if ( $verbose > CLI_NO ) {
         print_newline();
         console_write(STDOUT,'upgradingcompleted');
